+++ buession-redis/src/main/java/com/buession/redis/client/lettuce/operations/LettuceSentinelStringOperations.java
@@ -279,13 +279,13 @@ public final class LettuceSentinelStringOperations extends AbstractStringOperati
 	}
 
 	@Override
-	public String substr(final String key, final long start, final long end) {
+	public String substr(final String key, final int start, final int end) {
 		final CommandArguments args = CommandArguments.create("key", key).put("start", start).put("end", end);
 		return notCommand(client, ProtocolCommand.SUBSTR, args);
 	}
 
 	@Override
-	public byte[] substr(final byte[] key, final long start, final long end) {
+	public byte[] substr(final byte[] key, final int start, final int end) {
 		final CommandArguments args = CommandArguments.create("key", key).put("start", start).put("end", end);
 		return notCommand(client, ProtocolCommand.SUBSTR, args);
 	}
